+++ code/src/cljs/sixsq/nuvla/ui/utils/semantic_ui_extensions.cljs
@@ -1,7 +1,6 @@
 (ns sixsq.nuvla.ui.utils.semantic-ui-extensions
   (:require
     [re-frame.core :refer [subscribe]]
-    [reagent.core :as reagent]
     [reagent.core :as r]
     [sixsq.nuvla.ui.i18n.subs :as i18n-subs]
     [sixsq.nuvla.ui.utils.accordion :as accordion-utils]
@@ -81,8 +80,8 @@
 
 
 (defn Pagination
-  "Provide pagination element with more visible icons. Note: :totalitems is in lowercase not to interfere with
-   React DOM attributes."
+  "Provide pagination element with more visible icons. Note: :totalitems is in lowercase not to
+   interfere with React DOM attributes."
   [options]
   (let [tr (subscribe [::i18n-subs/tr])]
     [:div
@@ -92,10 +91,10 @@
                   :size  :medium}
         (str (@tr [:total]) ": " (:totalitems options))])
      [ui/Pagination
-      (merge {:firstItem {:content (reagent/as-element [ui/Icon {:name "angle double left"}]) :icon true}
-              :lastItem  {:content (reagent/as-element [ui/Icon {:name "angle double right"}]) :icon true}
-              :prevItem  {:content (reagent/as-element [ui/Icon {:name "angle left"}]) :icon true}
-              :nextItem  {:content (reagent/as-element [ui/Icon {:name "angle right"}]) :icon true}}
+      (merge {:firstItem {:content (r/as-element [ui/Icon {:name "angle double left"}]) :icon true}
+              :lastItem  {:content (r/as-element [ui/Icon {:name "angle double right"}]) :icon true}
+              :prevItem  {:content (r/as-element [ui/Icon {:name "angle left"}]) :icon true}
+              :nextItem  {:content (r/as-element [ui/Icon {:name "angle right"}]) :icon true}}
              (merge {:floated :right, :size "tiny"} options))]]))